long-sentence/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
nltk==3.8.1
python-docx==0.8.11
pdfplumber==0.9.0


